registration.py: Split out onboarding messages to onboarding.py.
